decode to int96, misc. fixes for deployment

+++ contracts/REXMarket.sol
@@ -610,7 +610,7 @@ abstract contract REXMarket is Ownable, SuperAppBase, Initializable {
                 distToken,
                 index,
                 subscriber,
-                shares, 
+                shares,
                 new bytes(0)
             ),
             new bytes(0), // user data
@@ -804,7 +804,7 @@ abstract contract REXMarket is Ownable, SuperAppBase, Initializable {
         (address _shareholder, int96 _flowRate,) = _getShareholderInfo(
             _agreementData, _superToken
         );
-        int _beforeFlowRate = abi.decode(_cbdata, (int));
+        int96 _beforeFlowRate = abi.decode(_cbdata, (int96));
 
 
         if (_shouldDistribute()) {
@@ -812,7 +812,7 @@ abstract contract REXMarket is Ownable, SuperAppBase, Initializable {
         }
 
         ShareholderUpdate memory _shareholderUpdate = ShareholderUpdate(
-          _shareholder, int96(_beforeFlowRate), _flowRate, _superToken
+          _shareholder, _beforeFlowRate, _flowRate, _superToken
         );
 
         // TODO: Udpate shareholder needs before and after flow rate
@@ -854,10 +854,10 @@ abstract contract REXMarket is Ownable, SuperAppBase, Initializable {
 
         _newCtx = _ctx;
         (address _shareholder, ) = abi.decode(_agreementData, (address, address));
-        (uint256 _uninvestAmount, int _beforeFlowRate ) = abi.decode(_cbdata, (uint256, int));
+        (uint256 _uninvestAmount, int96 _beforeFlowRate ) = abi.decode(_cbdata, (uint256, int96));
 
         ShareholderUpdate memory _shareholderUpdate = ShareholderUpdate(
-          _shareholder, int96(_beforeFlowRate), 0, _superToken
+          _shareholder, _beforeFlowRate, 0, _superToken
         );
 
         _newCtx = _updateShareholder(_newCtx, _shareholderUpdate);
